Remove a warning.

+++ Samples/Desktop/af/helper_dx12.cpp
@@ -54,7 +54,7 @@ void afWriteBuffer(const IBOID id, const void* buf, int size)
 	D3D12_RANGE readRange = {};
 	id->Map(0, &readRange, &p);
 	memcpy(p, buf, size);
-	D3D12_RANGE wroteRange = {0, size};
+	D3D12_RANGE wroteRange = {0, (SIZE_T)size};
 	id->Unmap(0, &wroteRange);
 }
